.agents/skills/reorganizar-topicos/scripts/pyproject.toml
[project]
name = "reorganizar-topicos-skill"
version = "0.1.0"
description = "Ambiente Python local da skill reorganizar-topicos"
requires-python = ">=3.12"
dependencies = [
  "pyyaml>=6.0",
]

[tool.uv]
package = false
